Simplified allpairs tests so that the worker is run in the foreground. This will help with reliability of the tests.

+++ allpairs/test/TR_allpairs_bitwise.sh
@@ -2,39 +2,44 @@
 
 . ../../dttools/src/test_runner.common.sh
 
-TEST_INPUT="./set.list"
-TEST_OUTPUT=bitwise.output
-PIDMASTER_FILE=allpairs.pid
-PIDWORKER_FILE=worker.pid
-PORT_FILE=WORKER.PORT
+TEST_INPUT=set.list
+TEST_OUTPUT=master.out
+PIDMASTER_FILE=master.pid
+PORT_FILE=worker.port
 
-prepare()
+export PATH=.:../src:../../work_queue/src:$PATH
+
+cleanfiles()
 {
 	rm -f $TEST_OUTPUT
 	rm -f $PIDMASTER_FILE
-	rm -f $PIDWORKER_FILE
 	rm -f $PORT_FILE
+	rm -f allpairs_multicore
 
-    ln -s ../src/allpairs_multicore .
+}
 
-    (PATH=.:$PATH ../src/allpairs_master -x 1 -y 1 --output-file $TEST_OUTPUT -Z $PORT_FILE $TEST_INPUT $TEST_INPUT BITWISE )&
-
-    pid=$!
-	echo $pid > $PIDMASTER_FILE
-
-	wait_for_file_creation $PORT_FILE 5
-	exit 0
+prepare()
+{
+	cleanfiles
 }
 
 run()
 {
-    ../../work_queue/src/work_queue_worker localhost `cat $PORT_FILE` --timeout 2 -d all
-    pid=$!
-	echo $pid > $PIDWORKER_FILE
+	ln -s ../src/allpairs_multicore .
 
-	wait_for_file_creation $TEST_OUTPUT 5
-	wait_for_file_modification $TEST_OUTPUT 3
+	echo "starting master"	
+	allpairs_master -x 1 -y 1 --output-file $TEST_OUTPUT -Z $PORT_FILE $TEST_INPUT $TEST_INPUT BITWISE &
 
+	pid=$!
+	echo $pid > $PIDMASTER_FILE
+
+	echo "waiting for $PORT_FILE to be created"
+	wait_for_file_creation $PORT_FILE 5
+
+	echo "starting worker"
+	work_queue_worker localhost `cat $PORT_FILE` --timeout 2
+
+	echo "checking output"
 	in_files=`cat "$TEST_INPUT" | awk -F"/" '{print $3}'`
 	howmany() { echo $#;}
 	num_files=$(howmany $in_files)
@@ -51,22 +56,19 @@ run()
 	  fi
 	done
 
+	echo "output is good"
 	exit 0
 
 }
 
 clean()
 {
-    kill -9 `cat $PIDMASTER_FILE`
-    kill -9 `cat $PIDWORKER_FILE`
-
-	rm -f $TEST_OUTPUT
-	rm -f $PIDMASTER_FILE
-	rm -f $PIDWORKER_FILE
-	rm -f $PORT_FILE
-    rm -f allpairs_multicore
-
-    exit 0
+	if [ -f master.pid ]
+	then
+		kill -9 `cat $PIDMASTER_FILE`
+	fi
+	cleanfiles
+	exit 0
 }
 
 dispatch $@
